# docs/swagger/openapi.yaml
openapi: "3.0.3"
info:
  title: Workflow Execution Service
  version: "1.0.0"
  description: >
    Runtime control plane for the BPMN-driven Workflow Engine. Manages workflow instances,
    task projections for dashboard queries, and user signals (claim, complete, defer, reassign).
    All identity and tenant context is injected by the Envoy API Gateway — callers never set
    these headers directly.
servers:
  - url: https://api.internal/execution
    description: Internal API gateway base (service prefix only; paths include /api/v1/)
security:
  - GatewayTenantContext: []
    UserContext: []
...
components:
  securitySchemes:
    GatewayTenantContext:
      type: apiKey
      in: header
      name: x-tenant-id
      description: Tenant UUID injected by the API Gateway.
    UserContext:
      type: apiKey
      in: header
      name: x-user-id
      description: Authenticated user UUID injected by the API Gateway.
    AdminRole:
      type: apiKey
      in: header
      name: x-tenant-roles
      description: >
        Comma-separated tenant roles injected by the API Gateway. Admin-only endpoints
        require at least one of: tenant_admin, tenant_owner.
    InternalServiceAuth:
      type: apiKey
      in: header
      name: x-internal-token
      description: >
        Service-to-service credential for the internal route group (RequireInternalToken,
        INTERNAL_API_TOKEN) — the same convention Definition Service already ships. These
        routes are NOT exposed through the Envoy public gateway and do NOT carry gateway
        identity headers; they are restricted to in-mesh callers by a Kubernetes
        NetworkPolicy AND authenticated by this shared secret — defense-in-depth, not either
        control alone (§9.2). The tenant context for these routes is taken from the request
        payload, not from x-tenant-id.
  parameters:
    TenantIdHeader:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant UUID (injected by gateway).
    UserIdHeader:
      name: x-user-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Authenticated user UUID (injected by gateway).
    TenantRolesHeader:
      name: x-tenant-roles
      in: header
      required: true
      schema:
        type: string
      description: >
        Comma-separated tenant roles. Must include tenant_admin or tenant_owner
        for admin-only endpoints.
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: >
        Optional client-supplied key for safe retry of this request. A replay with the same
        key and an identical request body returns the original cached response. A replay with
        the same key and a different body returns 409 IDEMPOTENCY_KEY_REPLAY.
    InstanceIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Workflow instance UUID.
...
    NodeKeyPath:
      name: node
      in: path
      required: true
      schema:
        type: string
      description: BPMN node key in the compiled plan whose assignee is being overridden.
      example: "review_stage_1"
...
  schemas:
...
    OverrideResult:
      type: object
      description: The persisted node-override record (rev 1.23 — matches overrideNodeAssignee's actual 200 response, not a SignalAccepted body).
      required:
        - workflow_instance_id
        - node_key
        - previous_user_id
        - new_user_id
        - actor_user_id
        - record_version
      properties:
        workflow_instance_id:
          type: string
          format: uuid
        node_key:
          type: string
          example: "review_stage_1"
        previous_user_id:
          type: string
          format: uuid
        new_user_id:
          type: string
          format: uuid
        reason:
          type: string
          nullable: true
        actor_user_id:
          type: string
          format: uuid
        record_version:
          type: integer
...
    ProblemDetails:
      type: object
      required: [type, title, status, detail, instance, code]
      properties:
        type:
          type: string
          format: uri
          description: >
            URI reference identifying the problem type.
            Format: https://errors.bcbp.io/execution/{kebab-code}
          example: "https://errors.bcbp.io/execution/task-concurrency-violation"
        title:
          type: string
          description: Short, human-readable problem title (invariant per error type).
          example: "Task Concurrency Violation"
        status:
          type: integer
          description: HTTP status code.
          example: 409
        detail:
          type: string
          description: Human-readable explanation of this specific occurrence.
          example: "task record_version mismatch; refresh and retry"
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying the specific request that produced the error.
          example: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/claim"
        code:
          type: string
          description: >
            Machine-readable error code (stable across releases). One of:
            BAD_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_ASSIGNEE, NOT_AUTHORIZED_FOR_RESOURCE,
            TENANT_MISMATCH, INSTANCE_NOT_FOUND, TASK_NOT_FOUND,
            TARGET_NODE_NOT_FOUND, OVERRIDE_NO_OP, DUPLICATE_BUSINESS_KEY, TENANT_NOT_ACTIVE,
            VERSION_NOT_PUBLISHED, VERSION_INVALID,
            RECORD_VERSION_CONFLICT, INSTANCE_ALREADY_TERMINAL, INVALID_INSTANCE_STATE,
            TASK_ALREADY_CLAIMED, CLAIM_NOT_APPLICABLE, INVALID_TASK_STATE,
            FORCE_BACK_NO_SAVED_BRANCH, NODE_ALREADY_RESOLVED, IDEMPOTENCY_KEY_REPLAY,
            ASSIGNEE_INELIGIBLE, OVERRIDE_MAP_INVALID, PAYLOAD_TOO_LARGE, UNSUPPORTED_MEDIA_TYPE,
            UPSTREAM_UNAVAILABLE, EVENT_DECODE_FAILED, INTERNAL_ERROR.
          example: "RECORD_VERSION_CONFLICT"
        invalid_params:
          type: array
          description: >
            Present on validation errors carrying more than one offending field/node — one entry
            per violation (e.g. ASSIGNEE_INELIGIBLE's bulk node-eligibility failures, §5.5).
            Absent (omitted entirely, not null) on every other error.
          items:
            type: object
            required: [name, reason]
            properties:
              name:
                type: string
                description: Name of the invalid field, query/path parameter, or node key.
                example: "review_finance"
              reason:
                type: string
                description: Human-readable explanation of why this entry is invalid.
                example: "user no longer holds required_level=SENIOR in department_id=9b1c...finance"
              code:
                type: string
                description: Optional machine-readable sub-code for this specific violation.
  responses:
...
    Unauthorized:
      description: Missing or invalid x-tenant-id / x-user-id gateway headers.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://errors.bcbp.io/execution/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "x-tenant-id header is required"
            instance: "/api/v1/instances"
            code: "UNAUTHORIZED"
    Forbidden:
      description: Caller lacks the required role or active assignment for this resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://errors.bcbp.io/execution/forbidden"
            title: "Forbidden"
            status: 403
            detail: "caller does not have an active assignment for this task"
            instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/claim"
            code: "NOT_ASSIGNEE"
    InstanceNotFound:
      description: Workflow instance not found or belongs to a different tenant.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://errors.bcbp.io/execution/instance-not-found"
            title: "Instance Not Found"
            status: 404
            detail: "workflow instance not found"
            instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6"
            code: "INSTANCE_NOT_FOUND"
...
    PayloadTooLarge:
      description: Request body exceeds the 10 MB limit (§5.10, §9.3 — same cap Definition Service applies).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://errors.bcbp.io/execution/payload-too-large"
            title: "Payload Too Large"
            status: 413
            detail: "request body exceeds 10 MB limit"
            instance: "/api/v1/instances"
            code: "PAYLOAD_TOO_LARGE"
    UnsupportedMediaType:
      description: Request Content-Type is not application/json.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://errors.bcbp.io/execution/unsupported-media-type"
            title: "Unsupported Media Type"
            status: 415
            detail: "request body must be application/json"
            instance: "/api/v1/instances"
            code: "UNSUPPORTED_MEDIA_TYPE"
    InternalError:
      description: Unhandled server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://errors.bcbp.io/execution/internal-error"
            title: "Internal Server Error"
            status: 500
            detail: "an unexpected error occurred"
            instance: "/api/v1/instances"
            code: "INTERNAL_ERROR"
    ServiceUnavailable:
      description: An upstream dependency (Temporal or Definition Service) is unavailable.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://errors.bcbp.io/execution/temporal-unavailable"
            title: "Service Unavailable"
            status: 503
            detail: "Temporal cluster is unreachable; retry after a short delay"
            instance: "/api/v1/instances"
            code: "UPSTREAM_UNAVAILABLE"
...
paths:
...
  /api/v1/instances/{id}/nodes/{node}/override:
    post:
      operationId: overrideNodeAssignee
      tags: [Instances]
      summary: Override the assignee for a specific node on a running instance (admin)
      description: >
        Admin override: replaces the default/current assignee for the given BPMN node with
        new_user_id, independent of the compiled plan's default assignment. Applies only to
        the node's current or future assignment; does not affect already-completed assignments.
        Returns 400 OVERRIDE_NO_OP if new_user_id is already the effective assignee for this node.
      security:
        - GatewayTenantContext: []
          UserContext: []
          AdminRole: []
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantRolesHeader"
        - $ref: "#/components/parameters/InstanceIdPath"
        - $ref: "#/components/parameters/NodeKeyPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_user_id, record_version]
              properties:
                new_user_id:
                  type: string
                  format: uuid
                  description: UUID of the user to assign this node to.
                reason:
                  type: string
                  minLength: 1
                  maxLength: 500
                  nullable: true
                  description: >
                    Optional human-readable reason for the override; stored on the
                    assignee_overrides audit record (§4.13).
                  example: "Original default assignee left the department"
                record_version:
                  type: integer
                  description: >
                    The caller's current view of the task's record_version at this node,
                    checked before the eligibility call and again before persisting (§5.4).
                  example: 4
      responses:
        "200":
          description: Override persisted and signaled; returns the override record (rev 1.23 — not a 202/SignalAccepted body).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverrideResult"
        "400":
          description: >
            Malformed request body, or code field OVERRIDE_NO_OP — new_user_id is already
            the effective assignee for this node.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                BadRequest:
                  value:
                    type: "https://errors.bcbp.io/execution/bad-request"
                    title: "Bad Request"
                    status: 400
                    detail: "field 'new_user_id' is required"
                    instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/nodes/review_stage_1/override"
                    code: "BAD_REQUEST"
                OverrideNoOp:
                  value:
                    type: "https://errors.bcbp.io/execution/override-no-op"
                    title: "Override No-Op"
                    status: 400
                    detail: "new_user_id is already the effective assignee for node 'review_stage_1'"
                    instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/nodes/review_stage_1/override"
                    code: "OVERRIDE_NO_OP"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/InstanceNotFound"
        "409":
          description: >
            State conflict. See code field: NODE_ALREADY_RESOLVED — the node has already
            completed or advanced past; RECORD_VERSION_CONFLICT — a concurrent change raced
            this request; IDEMPOTENCY_KEY_REPLAY — same key reused with a different request body.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                NodeAlreadyResolved:
                  value:
                    type: "https://errors.bcbp.io/execution/node-already-resolved"
                    title: "Node Already Resolved"
                    status: 409
                    detail: "node 'review_stage_1' has already been completed or advanced past"
                    instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/nodes/review_stage_1/override"
                    code: "NODE_ALREADY_RESOLVED"
                RecordVersionConflict:
                  value:
                    type: "https://errors.bcbp.io/execution/record-version-conflict"
                    title: "Record Version Conflict"
                    status: 409
                    detail: "a concurrent change was applied to this node; refresh and retry"
                    instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/nodes/review_stage_1/override"
                    code: "RECORD_VERSION_CONFLICT"
                IdempotencyReplay:
                  value:
                    type: "https://errors.bcbp.io/execution/idempotency-key-replay"
                    title: "Idempotency Key Replay"
                    status: 409
                    detail: "Idempotency-Key reused with a different request body"
                    instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/nodes/review_stage_1/override"
                    code: "IDEMPOTENCY_KEY_REPLAY"
        "422":
          description: The specified new_user_id is not an eligible assignee for this node.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://errors.bcbp.io/execution/assignee-ineligible"
                title: "Assignee Ineligible"
                status: 422
                detail: "new_user_id is not eligible to be assigned to node 'review_stage_1'"
                instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/nodes/review_stage_1/override"
                code: "ASSIGNEE_INELIGIBLE"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...